+++ packages/terra-status/src/Status.jsx
@@ -3,6 +3,7 @@ import PropTypes from 'prop-types';
 import classNames from 'classnames/bind';
 import 'terra-base/lib/baseStyles';
 import styles from './Status.module.scss';
+import VisuallyHiddenText from '../../terra-visually-hidden-text'
 
 const cx = classNames.bind(styles);
 
@@ -15,10 +16,12 @@ const propTypes = {
    * The color of the status indicator. Accepts a CSS color value.
    */
   color: PropTypes.string,
+  visuallyHiddenText: PropTypes.string,
 };
 
-const Status = ({ color, children, ...customProps }) => (
+const Status = ({ color, children, visuallyHiddenText, ...customProps }) => (
   <div {...customProps} style={{ borderColor: color }} className={cx('status', customProps.className)}>
+    <VisuallyHiddenText text={visuallyHiddenText} />
     {children}
   </div>
 );