+++ packages/terra-status/src/terra-dev-site/test/status/StatusDefault.test.jsx
@@ -5,7 +5,7 @@ const simpleText = <div style={{ padding: '5px' }}>Sample text</div>;
 
 const StatusDefault = () => (
   <div>
-    <Status id="status" color="#ff0000">{simpleText}</Status>
+    <Status id="status" color="#ff0000" visuallyHiddenText="This is a test.">{simpleText}</Status>
   </div>
 );